+++ packages/access-definitions/src/Wrapping/Utilities/SchemaPathProcessor.php
@@ -44,7 +44,7 @@ class SchemaPathProcessor
     /**
      * Check the paths of the given conditions for availability and applies aliases using the given type.
      *
-     * Also adds the {@link ReadableTypeInterface::getAccessCondition() access condition} of the given type.
+     * Also adds the {@link ReadableTypeInterface::getAccessConditions() access conditions} of the given type.
      *
      * @template TCondition of \EDT\Querying\Contracts\PathsBasedInterface
      * @template TSorting of \EDT\Querying\Contracts\PathsBasedInterface
@@ -57,7 +57,7 @@ class SchemaPathProcessor
      * @throws PathException Thrown if {@link TypeInterface::getAliases()} returned an invalid path.
      * @throws AccessException
      *
-     * @deprecated use {@link SchemaPathProcessor::mapFilterConditions()} and {@link SchemaPathProcessor::processAccessCondition()} instead
+     * @deprecated use {@link SchemaPathProcessor::mapFilterConditions()} and {@link SchemaPathProcessor::processAccessConditions()} instead
      */
     public function mapConditions(TypeInterface $type, PathsBasedInterface ...$conditions): array
     {
@@ -75,9 +75,10 @@ class SchemaPathProcessor
         }
 
         // process the access condition too, however based on a different property set than the external conditions
-        $conditions[] = $this->processAccessCondition($type);
-
-        return $conditions;
+        return array_merge(
+            $conditions,
+            $this->processAccessConditions($type)
+        );
     }
 
     /**
@@ -205,7 +206,7 @@ class SchemaPathProcessor
     }
 
     /**
-     * Get the processed {@link ReadableTypeInterface::getAccessCondition() access condition}
+     * Get the processed {@link ReadableTypeInterface::getAccessConditions() access conditions}
      * of the given type.
      *
      * @template TCondition of \EDT\Querying\Contracts\PathsBasedInterface
@@ -213,20 +214,22 @@ class SchemaPathProcessor
      *
      * @param TypeInterface<TCondition, TSorting, object> $type
      *
-     * @return TCondition
+     * @return list<TCondition>
      *
      * @throws AccessException
      * @throws PathException
      *
      * @internal
      */
-    public function processAccessCondition(TypeInterface $type): PathsBasedInterface
+    public function processAccessConditions(TypeInterface $type): array
     {
-        $condition = $type->getAccessCondition();
+        $conditions = $type->getAccessConditions();
         $typeAccessor = new InternTypeAccessor($this->typeProvider);
         $processor = $this->propertyPathProcessorFactory->createPropertyPathProcessor($typeAccessor);
-        $processor->processPropertyPaths($condition, $type);
+        array_map(static function (PathsBasedInterface $condition) use ($processor, $type): void {
+            $processor->processPropertyPaths($condition, $type);
+        }, $conditions);
 
-        return $condition;
+        return $conditions;
     }
 }